+++ OpenGL/OpenGL/src/Texture.cpp
@@ -31,7 +31,7 @@ Texture::Texture(const std::string& path)
 	}
 }
 
-Texture::Texture(const std::string& path, std::map<GLenum, GLfloat> options)
+Texture::Texture(const std::string& path, std::map<GLenum, GLint> options)
 	: m_Texture(0), m_FilePath(path), m_LocalBuffer(nullptr), m_Width(0), m_Height(0), m_BPP(0)
 {
 	if (options.size() != 4)
